gdb: Replace long() with int()
Python 3 doesn't have 'long' anymore, so commands using it fail with
newer GDB. long on python2 is the same as int on python3, both are
arbitrary-precision. On python2 int is fixed-precision, but seems to
be still enough (64 bit), so use that instead.

Message-Id: <1531215600-31899-1-git-send-email-tgrabiec@scylladb.com>

+++ scylla-gdb.py
@@ -741,7 +741,7 @@ def get_seastar_memory_start_and_size():
     cpu_mem = gdb.parse_and_eval('\'seastar::memory::cpu_mem\'')
     page_size = int(gdb.parse_and_eval('\'seastar::memory::page_size\''))
     total_mem = int(cpu_mem['nr_pages']) * page_size
-    start = long(cpu_mem['memory'])
+    start = int(cpu_mem['memory'])
     return start, total_mem
 
 def seastar_memory_layout():
@@ -761,7 +761,7 @@ class scylla_ptr(gdb.Command):
     def __init__(self):
         gdb.Command.__init__(self, 'scylla ptr', gdb.COMMAND_USER, gdb.COMPLETE_COMMAND)
     def invoke(self, arg, from_tty):
-        ptr = long(arg, 0)
+        ptr = int(arg, 0)
 
         owning_thread = None
         for t, start, size in seastar_memory_layout():
@@ -779,7 +779,7 @@ class scylla_ptr(gdb.Command):
 
         cpu_mem = gdb.parse_and_eval('\'seastar::memory::cpu_mem\'')
         page_size = int(gdb.parse_and_eval('\'seastar::memory::page_size\''))
-        offset = ptr - long(cpu_mem['memory'])
+        offset = ptr - int(cpu_mem['memory'])
         ptr_page_idx = offset / page_size
         pages = cpu_mem['pages']
         page = pages[ptr_page_idx];
@@ -1309,7 +1309,7 @@ def find_in_live(mem_start, mem_size, value, size_selector='g'):
             if 'live' in ptr_info:
                 m = re.search('live \((0x[0-9a-f]+)', ptr_info)
                 if m:
-                    obj_start = long(m.group(1), 0)
+                    obj_start = int(m.group(1), 0)
                     addr = int(line, 0)
                     offset = addr - obj_start
                     yield obj_start, offset
